Users can join others activities and get put into chats automatically
When users click join in an activity it adds them to a chat and brings them to messages. However you need to refresh the messages page for the new chatt which they just joined to appear, and the navbar tweeks out with the link to messages for some reason and once you click join you cant go back to activities.